Fixed the ridirection to dashboards
